refactor(mcp): code-review pass — settings, naming, DI, coding style
Per-review changes:
- McpSettings: drop IsEnabled/ServerName/ServerVersion. Empty/null Route
  now means "disabled" (commented above the property). Both module
  registration and endpoint mapping respect this.
- Server name hard-coded to "Voxt" (the product name); version uses
  ApiConstants.VersionString — same as RPC's ServerApiInfo comparison
  key. The 401 WWW-Authenticate realm now matches: "Voxt".
- Tool classes (ChatTools, MessageTools) now take IServiceProvider and
  resolve dependencies via properties — matches the Chats(services)
  pattern in CODING_STYLE.md. Added a private `Session` computed
  property so methods drop the per-call `var session = ...` line.
- Renamed: ChatTools (message ops) -> MessageTools;
  DiscoveryTools (chat listing) -> ChatTools. Tests renamed via git mv
  to preserve history.
- Added "// Private methods" section separators per coding style.

All 18 integration tests still pass.
