+++ src/dotnet/Mcp/Module/McpSettings.cs
@@ -2,8 +2,6 @@ namespace ActualChat.Mcp.Module;
 
 public sealed class McpSettings
 {
-    public bool IsEnabled { get; set; } = true;
+    // Empty/null disables the MCP server entirely.
     public string Route { get; set; } = "/api/mcp";
-    public string ServerName { get; set; } = "ActualChat";
-    public string ServerVersion { get; set; } = "1.0.0";
 }
